Week5/vlm_project/requirements.txt
# Week 5: VLM Project
# Install: pip install -r requirements.txt
# (All packages should already be in the llm-learning conda env)

torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
datasets>=2.16.0
Pillow>=9.0
matplotlib>=3.6
numpy>=1.24
pyyaml>=6.0
tqdm>=4.65
